+++ CHANGELOG.md
@@ -11,6 +11,21 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-beta.67 (2019-09-07)
+
+#### :rocket: New Feature
+
+* Improved number prelude API:
+  * Added `Number.isFloat`
+  * Added `Number.isEven`
+  * Added `Number.isOdd`
+  * Added `Number.isPositive`
+  * Added `Number.isNegative`
+  * Added `Number.isNonNegative`
+  * Added `Number.isNatural`
+  * Added `Number.isBetweenZeroAndOne`
+  * Added `Number.isPositiveBetweenZeroAndOne`
+
 ## v3.0.0-beta.66 (2019-08-27)
 
 #### :bug: Bug Fix